feat(contact-tagging): Moved pages/admin/callouts/ToggleTagButton.vue -> tag/ToggleTagButton.vue
