+++ apps/frontend/src/components/tag/ToggleTagButton.vue
@@ -29,8 +29,8 @@
 <script lang="ts" setup>
 import { faCog, faTag } from '@fortawesome/free-solid-svg-icons';
 import { useI18n } from 'vue-i18n';
-import AppSelectableList from '../../../AppSelectableList.vue';
-import AppDropdownButton from '../../../button/AppDropdownButton.vue';
+import AppSelectableList from '@components/AppSelectableList.vue';
+import AppDropdownButton from '@components/button/AppDropdownButton.vue';
 
 const emit = defineEmits<{
   (event: 'toggle', id: string, successText: string): void;
